requirements.txt
# HTTP and HTML handling
requests>=2.25.1
beautifulsoup4>=4.9.3
urllib3>=1.26.5
lxml==4.9.3

# PDF processing
PyPDF2>=2.0.0

# Date parsing
python-dateutil>=2.8.1

# Progress bars
tqdm>=4.61.1

# Configuration
pyyaml>=6.0

# Caching
diskcache>=5.2.1

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Type checking
mypy>=0.910
types-requests>=2.25.0
types-PyYAML>=6.0.0
types-python-dateutil>=2.8.0
types-beautifulsoup4==4.12.0.6

# Logging
colorlog>=6.6.0

# CLI
click>=8.0.1

# Data processing
pandas>=1.3.0
